[tests] Factor out reusable components
- Move the random number generator into src/random.rs
- Move BlockVisitor and Stamper into src/io_engine/test_utils.rs
