+++ src/io_engine/mod.rs
@@ -22,3 +22,6 @@ pub mod core;
 
 #[cfg(test)]
 pub mod ramdisk;
+
+#[cfg(any(test, feature = "devtools"))]
+pub mod test_utils;
